api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
scikit-learn
joblib
numpy
pandas
gradio-client
google-genai
pydantic_settings
Pillow
huggingface-hub
python-dotenv
opencv-contrib-python
mediapipe
torch
torchvision
onnxruntime